refactor: 선호 장르 컬럼 추가 (회원가입/조회/수정)

+++ src/main/java/com/booksy/domain/user/dto/SignupRequest.java
@@ -21,5 +21,5 @@ public class SignupRequest {
   private String nickname;  // nullable → email로 대체
   private String profileImage;  // nullable
 
-  private List<Integer> preferredCategoryIds;  // nullable
+  private List<Long> preferredCategoryIds;
 }
